+++ src/app/components/Registro/formularioRegistro.tsx
@@ -9,8 +9,9 @@ import React, { useState } from "react";
 import { InsertarRegistro } from "./insertarUsuarioRegistro";
 import { FormRegistroValidation } from "../../../../tipos/tiposForm";
 import { useRouter } from "next/navigation";
-import { Toaster, toast } from 'sonner'
-import { setCookie } from "cookies-next"
+import { Toaster, toast } from "sonner";
+import { setCookie } from "cookies-next";
+import { InsertarCarritoExistente } from "./insertarCarritoLocal";
 
 export default function FormRegistro() {
   const [showPassword, setShowPassword] = useState<boolean>(false);
@@ -33,7 +34,7 @@ export default function FormRegistro() {
     const result = FormRegistroValidation.safeParse(newForm);
     if (!result.success) {
       setIsLoading(false);
-      toast.error(result.error.issues[0].message)
+      toast.error(result.error.issues[0].message);
       return;
     }
 
@@ -41,11 +42,17 @@ export default function FormRegistro() {
     const response = await InsertarRegistro(result.data);
     if (response?.error) {
       //Manejar el error
-      toast.error(response.error)
+      toast.error(response.error);
     } else {
-      localStorage.clear()
+      //Chequeamos si el cliente tiene productos en el carrito
+      const carritoString = localStorage.getItem("carrito");
+      if (carritoString !== null) {
+        const carritoObjeto = JSON.parse(carritoString);
+        await InsertarCarritoExistente(carritoObjeto, result.data.correoElectronico)
+        localStorage.clear();
+      }
       //Seteamos la cookie con el token que contiene la informacion del usuario
-      setCookie("client-Token", response.token)
+      setCookie("client-Token", response.token);
 
       //Si se ha insertado correctamente redirige al perfil
       router.push("/Perfil");
@@ -54,7 +61,7 @@ export default function FormRegistro() {
   };
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault(); 
+    event.preventDefault();
     if (isLoading) return;
     const formData = new FormData(event.currentTarget);
     await clientAction(formData);
@@ -133,7 +140,8 @@ export default function FormRegistro() {
       <button
         type="submit"
         className=" w-2/3 bg-colorBase h-14 self-center text-2xl hover:bg-colorBaseSecundario hover:text-black transition duration-300 ease-in-out"
-        disabled={isLoading}>
+        disabled={isLoading}
+      >
         Crear cuenta
       </button>
       <Toaster position="top-right" richColors />
